requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
cffi==1.15.1
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-MySQLdb==1.0.1
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
pycparser==2.21
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1
